just one bit

+++ code/procs/gamehelpers.dm
@@ -106,7 +106,7 @@ var/list/stinkThingies = list("ass","taint","armpit","excretions","leftovers","a
 	else
 		if (iscarbon(user))
 			var/mob/living/carbon/C = user
-			if (C.bioHolder.HasEffect("telekinesis") && get_dist(source, user) <= 7) //You can only reach stuff within your screen.
+			if ((C.bioHolder.HasEffect("telekinesis")) && get_dist(source, user) <= 7) //You can only reach stuff within your screen.
 				var/X = source:x
 				var/Y = source:y
 				var/Z = source:z
@@ -124,6 +124,34 @@ var/list/stinkThingies = list("ass","taint","armpit","excretions","leftovers","a
 					O.icon = 'icons/effects/effects.dmi'
 					O.icon_state = "nothing"
 					flick("empdisable",O)
+					sleep(0.5 SECONDS)
+					qdel(O)
+
+				return TRUE
+
+			else if (istype(C.organHolder.augmentation_nerve, /obj/item/organ/augmentation/head/wireless_interact) && get_dist(source, user) <= 7)
+				var/obj/item/organ/augmentation/head/wireless_interact/WI = C.organHolder.augmentation_nerve
+				var/X = source:x
+				var/Y = source:y
+				var/Z = source:z
+				if(WI.flashed == TRUE)
+					boutput(user, "<span class='alert'>Your augmentation is recovering from being flashed.</span>")
+					return FALSE
+				if(isrestrictedz(Z) || isrestrictedz(user.z))
+					boutput(user, "<span class='alert'>Your augmentation doesn't seem to work here.</span>")
+					return FALSE
+				if(!istype(source, /obj) || istype(source, /obj/item))
+					return FALSE
+				SPAWN_DBG(0)
+					var/obj/overlay/O = new /obj/overlay ( locate(X,Y,Z) )
+					O.name = "magnetic touch"
+					O.anchored = 1
+					O.set_density(0)
+					O.layer = FLY_LAYER
+					O.set_dir(pick(cardinal))
+					O.icon = 'icons/effects/effects.dmi'
+					O.icon_state = "nothing"
+					flick("emppulse",O)
 					sleep(0.5 SECONDS)
 					qdel(O)
 
@@ -686,3 +714,9 @@ proc/ThrowRandom(var/atom/movable/A, var/dist = 10, var/speed = 1, var/list/para
 	if(istype(A))
 		var/turf/Y = GetRandomPerimeterTurf(A, dist)
 		A.throw_at(Y, dist, speed, params, thrown_from, throw_type, allow_anchored, bonus_throwforce, end_throw_callback)
+
+proc/HasWirelessInteractor(var/mob/living/carbon/human/H)
+	if(H.organHolder && istype(H.organHolder.augmentation_nerve, /obj/item/organ/augmentation/head/wireless_interact))
+		return TRUE
+	else
+		return FALSE
